client: make local development possible
Day 6 of trying to break the curse placed on my laptop has failed, and
internet has now degraded to simply never working on WiFi. It turns out
that copying the code to the repo is the only sane way of doing it, and
I regret not making sure we did this earlier.

We can't do it for the material icons /trivially/ because the CSS
contains references to online fonts which we would have to edit the CSS
which will be very fun.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>